fix: return notAvailable if not enabled
The http routes are registered even if the plugin is not
enabled, so we need to return something in case somebody
knocks on the door. Previously these http requests were
left dangling, causing a browser to run out of sockets
if for example vesselpositions was there trying to
retrieve tracks.